backend-python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.2.6
langchain-google-genai==1.0.6
langchain-community==0.2.6
chromadb==0.5.0
pypdf2==3.0.1
python-docx==1.1.2
pydantic==2.7.4
google-generativeai==0.7.2
httpx==0.27.0